Consolidated and updated Product page and changed routes to the page
