+++ src/pages/product/Product.scss
@@ -21,6 +21,15 @@
       width: 70%;
       min-width: 300px;
       max-width: 650px;
+      padding: 1.5rem 0;
+      display: flex;
+      flex-direction: column;
+      gap: 1.2rem;
+
+      .optionButtons {
+        display: flex;
+        gap: 1rem;
+      }
 
       .transactionForm {
         width: 100%;
